btnClick calls playGame(); Commented out humanChoice.
After testing this addition, I find that the game will only get the
user's choice once, and this choice is used all through the 5 rounds.
This is not the intended outcome. Need to find a way for playGame() to
wait for the user's (input), this time button clicks before proceeding.